refactor pdp.py script components into separate files
The pdp.py script has now been refactored further into separate
pdp_script.py, subcommands.py, tools.py, and parsers.py modules
